Add integration tests for consensus service
